refactor: move skill operations to SDK with AgentResolver DI (Phase 5)
Move 11 skill-*.ts files from CLI lib/ to SDK. Agent-dependent modules
accept an optional AgentResolver parameter. CLI shims re-export with
the CLI agent resolver auto-injected for backward compatibility.
