+++ packages/cli/src/lib/skill-filters.ts
@@ -1,96 +1,21 @@
-/**
- * Filter utilities for skill listings.
- *
- * Provides reusable filter functions and CLI argument definitions for
- * `--agent` and `--skill` flags used across skill list, add, and remove
- * commands.
- */
+// Re-export shim — moved to @agents/sdk/providers/local/skill/filters
+// The SDK version's validateAgentFilter takes an optional AgentResolver param;
+// the CLI shim wraps it to inject the CLI agent resolver automatically.
 
+export {
+  filterArgs,
+  filterByAgent,
+  filterBySkill,
+  type InstalledSkillEntry,
+} from '@agents/sdk/providers/local/skill/filters'
+
+import { validateAgentFilter as sdkValidateAgentFilter } from '@agents/sdk/providers/local/skill/filters'
 import type { OutputFormatter } from '@agents/sdk/ui'
-import type { AgentType } from './agents'
-import { AGENT_CONFIGS } from './agents'
-
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
-export interface InstalledSkillEntry {
-  /** Skill name (from frontmatter). */
-  name: string
-  /** Agents that have this skill installed. */
-  agents: AgentType[]
-  /** Source string (e.g. "owner/repo" or "./local/path"). */
-  source?: string
-  /** SHA256 hex digest of the skill directory. */
-  computedHash?: string
-  /** Symlink target path (if installed via symlink). */
-  symlinkTarget?: string
-}
-
-// ---------------------------------------------------------------------------
-// Filter functions
-// ---------------------------------------------------------------------------
+import { createCliAgentResolver } from './agents'
 
 /**
- * Filter skills to only those installed for a specific agent.
- *
- * Returns all skills when `agentFilter` is undefined or empty.
- */
-export function filterByAgent(
-  skills: InstalledSkillEntry[],
-  agentFilter?: string
-): InstalledSkillEntry[] {
-  if (!agentFilter) return skills
-  return skills.filter((s) => s.agents.includes(agentFilter as AgentType))
-}
-
-/**
- * Filter skills to only those matching a specific name (case-insensitive).
- *
- * Returns all skills when `skillFilter` is undefined or empty.
- */
-export function filterBySkill(
-  skills: InstalledSkillEntry[],
-  skillFilter?: string
-): InstalledSkillEntry[] {
-  if (!skillFilter) return skills
-  const lower = skillFilter.toLowerCase()
-  return skills.filter((s) => s.name.toLowerCase() === lower)
-}
-
-/**
- * Validate that an agent name is known. Emits a warning via the output
- * formatter if the agent is not in the registry.
- *
- * @returns `true` if the agent is valid, `false` otherwise.
+ * CLI-specific validateAgentFilter that auto-injects the CLI agent resolver.
  */
 export function validateAgentFilter(agentName: string, out: OutputFormatter): boolean {
-  if (AGENT_CONFIGS.has(agentName as AgentType)) return true
-  const known = [...AGENT_CONFIGS.keys()].slice(0, 5)
-  out.warn(`Unknown agent "${agentName}". Known agents: ${known.join(', ')}...`)
-  return false
+  return sdkValidateAgentFilter(agentName, out, createCliAgentResolver())
 }
-
-// ---------------------------------------------------------------------------
-// Shared CLI argument definitions
-// ---------------------------------------------------------------------------
-
-/**
- * Reusable citty argument definitions for `--agent` and `--skill` flags.
- *
- * Usage:
- * ```ts
- * import { filterArgs } from './skill-filters'
- *
- * defineCommand({
- *   args: { ...filterArgs, ... },
- * })
- * ```
- */
-export const filterArgs = {
-  agent: {
-    type: 'string' as const,
-    description: 'Filter by agent name (e.g., claude-code, cursor)',
-  },
-  skill: { type: 'string' as const, description: 'Filter by skill name' },
-} as const
